# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Schedule API",
        "description": "API для работы с функционалом раписаний занятий в Университетах",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/admin/settings": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Получение списка режимов звонков",
                "operationId": "get_settings",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Settings"
                ],
                "summary": "Создание режима звонков",
                "operationId": "add_setting",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/admin/settings/{settingId}": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Получение одного режима звонков",
                "operationId": "show_setting",
                "parameters": [
                    {
                        "name": "settingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Settings"
                ],
                "summary": "Обновление режима звонков",
                "operationId": "edit_setting",
                "parameters": [
                    {
                        "name": "settingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "setting_items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "time_start": {
                                                    "type": "string"
                                                },
                                                "time_end": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Settings"
                ],
                "summary": "Удаление режима звонков",
                "operationId": "delete_setting",
                "parameters": [
                    {
                        "name": "settingId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Enter token in string format!",
                "name": "token",
                "in": "header"
            },
            "x_account_id": {
                "type": "apiKey",
                "description": "iDesk account ID",
                "name": "X-Account-Id",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "token": [
                "get_settings",
                "add_setting",
                "edit_setting",
                "show_setting",
                "delete_setting"
            ],
            "x_account_id": [
                "get_settings",
                "add_setting",
                "edit_setting",
                "show_setting",
                "delete_setting"
            ]
        }
    ]
}